refactor warnings to utils/xml
